regroup api types into one module and rename 'Cardano.Wallet.Api' to 'Cardano.Api'
