+++ cardano-wallet.cabal
@@ -62,6 +62,8 @@ library
   hs-source-dirs:
       src
   exposed-modules:
+      Cardano.Api
+      Cardano.Api.Types
       Cardano.DBLayer
       Cardano.DBLayer.MVar
       Cardano.NetworkLayer
@@ -70,26 +72,11 @@ library
       Cardano.Wallet
       Cardano.Wallet.AddressDerivation
       Cardano.Wallet.AddressDiscovery
-      Cardano.Wallet.Api
-      Cardano.Wallet.Api.JSON
-      Cardano.Wallet.Api.Types.Amount
-      Cardano.Wallet.Api.Types.CurrencyUnit
-      Cardano.Wallet.Api.Types.Percentage
-      Cardano.Wallet.Api.Types.Wallet
-      Cardano.Wallet.Api.Types.WalletAddressPoolGap
-      Cardano.Wallet.Api.Types.WalletBalance
-      Cardano.Wallet.Api.Types.WalletDelegation
-      Cardano.Wallet.Api.Types.WalletDelegationStatus
-      Cardano.Wallet.Api.Types.WalletId
-      Cardano.Wallet.Api.Types.WalletName
-      Cardano.Wallet.Api.Types.WalletPassphraseInfo
-      Cardano.Wallet.Api.Types.WalletState
-      Cardano.Wallet.Api.Types.WalletStateStatus
       Cardano.Wallet.Binary
       Cardano.Wallet.Binary.Packfile
-      Cardano.WalletLayer
       Cardano.Wallet.Mnemonic
       Cardano.Wallet.Primitive
+      Cardano.WalletLayer
       Servant.Extra.ContentTypes
   other-modules:
       Paths_cardano_wallet
@@ -140,6 +127,7 @@ test-suite unit
       Main.hs
   other-modules:
       Cardano.DBLayer.MVarSpec
+      Cardano.Api.TypesSpec
       Cardano.DBLayerSpec
       Cardano.Launcher
       Cardano.LauncherSpec
@@ -148,8 +136,6 @@ test-suite unit
       Cardano.NetworkLayerSpec
       Cardano.Wallet.AddressDerivationSpec
       Cardano.Wallet.AddressDiscoverySpec
-      Cardano.Wallet.Api.Arbitraries
-      Cardano.Wallet.Api.JsonSpec
       Cardano.Wallet.Binary.PackfileSpec
       Cardano.Wallet.BinarySpec
       Cardano.Wallet.MnemonicSpec
