--- src/Cardano/Wallet/Api/Types/WalletPassphraseInfo.hs
@@ -1,23 +1,0 @@
-{-# LANGUAGE DeriveGeneric #-}
-
-module Cardano.Wallet.Api.Types.WalletPassphraseInfo where
-
-import Prelude
-
-import Cardano.Wallet.Api.JSON
-    ( defaultRecordTypeOptions )
-import Data.Aeson
-    ( FromJSON (..), ToJSON (..), genericParseJSON, genericToJSON )
-import Data.Time.Clock
-    ( UTCTime )
-import GHC.Generics
-    ( Generic )
-
-newtype WalletPassphraseInfo = WalletPassphraseInfo
-    { _lastUpdatedAt :: UTCTime
-    } deriving (Eq, Generic, Show)
-
-instance FromJSON WalletPassphraseInfo where
-    parseJSON = genericParseJSON defaultRecordTypeOptions
-instance ToJSON WalletPassphraseInfo where
-    toJSON = genericToJSON defaultRecordTypeOptions
